pyproject.toml
[project]
name = "tn_agent_launcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.15",
    "channels==4.2.*",
    "channels-redis==4.2",
    "daphne>=4.1.2",
    "dj-database-url==0.5.0",
    "dj-rest-auth==7.0.*",
    "django==4.2.*",
    "django-anymail==8.4",
    "django-cors-headers==3.8.0",
    "django-currentuser==0.7.0",
    "django-extensions==3.2.*",
    "django-filter==2.4.*",
    "django-model-utils==4.1.1",
    "django-multiselectfield==0.1.12",
    "django-storages==1.11.1",
    "django4-background-tasks>=1.2.10",
    "djangorestframework==3.15.*",
    "drf-nested-routers==0.93.3",
    "drf-spectacular>=0.28.0",
    "drf-writable-nested==0.6.3",
    "gunicorn>=23.0.0",
    "openai>=1.65.0",
    "pandas>=2.3.3",
    "pre-commit==2.13.*",
    "premailer>=3.10.0",
    "psycopg2-binary==2.9.*",
    "pydantic>=2.10.0",
    "pydantic-ai>=0.2.20",
    "pydantic-ai-bedrock>=0.1.0",
    "pytest-cov>=6.0.0",
    "pytest-django==4.5.*",
    "pytest-factoryboy==2.7.*",
    "pytest-mccabe>=2.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0",
    "python-slugify==5.0.2",
    "pytz==2021.1",
    "requests==2.*",
    "rollbar>=1.2.0",
    "twilio>=9.4.6",
    "whitenoise==5.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.8",
    "pytest==7.1.*",
    "factory-boy==3.2.*",
    "debugpy>=1.8.16",
]
lambda-dev = [
    # Lambda-specific development dependencies
    "pytest-asyncio>=0.21.0,<0.24.0",  # Compatible with pytest 7.x
    "mypy>=1.0.0",
    "types-boto3>=1.0.0",
]

[project.optional-dependencies]
doc-processing = [
    # Document processing dependencies - not available on Heroku
    "docling>=2.48.0",
]

[tool.ruff]
line-length = 100
exclude = [
    "**/migrations/*"
]


[tool.ruff.lint]
extend-select = ["I"]
